move channel manager into its mod.rs and fix visibility of methods
