+++ miner-apps/translator/src/lib/sv2/mod.rs
@@ -1,5 +1,5 @@
 pub mod channel_manager;
 pub mod upstream;
 
-pub use channel_manager::channel_manager::ChannelManager;
+pub use channel_manager::ChannelManager;
 pub use upstream::upstream::Upstream;
